renames shapes/ to elements/ for now
