+++ src/app/patterns/body/explorer/CompareOverTimeGraph.svelte
@@ -7,7 +7,7 @@ import LeftAxis from '../../../../components/data-graphics/guides/LeftAxis.svelt
 import BottomAxis from '../../../../components/data-graphics/guides/BottomAxis.svelte';
 import GraphicBody from '../../../../components/data-graphics/GraphicBody.svelte';
 import BuildIDRollover from '../../../../components/data-graphics/rollovers/BuildIDRollover.svelte';
-import Line from '../../../../components/data-graphics/shapes/Line.svelte';
+import Line from '../../../../components/data-graphics/elements/Line.svelte';
 import ReferenceSymbol from '../elements/ReferenceSymbol.svelte';
 
 import { cartesianCoordSpring } from '../utils/animation';
