+++ src/app/patterns/body/explorer/DistributionComparison.svelte
@@ -6,7 +6,7 @@ import TopAxis from '../../../../components/data-graphics/guides/TopAxis.svelte'
 import BottomAxis from '../../../../components/data-graphics/guides/BottomAxis.svelte';
 import RightAxis from '../../../../components/data-graphics/guides/RightAxis.svelte';
 import ReferenceSymbol from '../elements/ReferenceSymbol.svelte';
-import Line from '../../../../components/data-graphics/shapes/Line.svelte';
+import Line from '../../../../components/data-graphics/elements/Line.svelte';
 
 import { nearestBelow } from '../../../../utils/stats';
