+++ src/app/patterns/body/table/ProportionComparisonCell.svelte
@@ -3,7 +3,7 @@ import { spring, tweened } from 'svelte/motion';
 import { cubicOut as easing } from 'svelte/easing';
 import DataGraphic from '../../../../components/data-graphics/DataGraphic.svelte';
 import TopAxis from '../../../../components/data-graphics/guides/TopAxis.svelte';
-import Violin from '../../../../components/data-graphics/shapes/Violin.svelte';
+import Violin from '../../../../components/data-graphics/elements/Violin.svelte';
 import ReferenceSymbol from '../elements/ReferenceSymbol.svelte';
 
 import { comparisonSmallMultiple } from '../utils/constants';
