+++ src/components/data-graphics/elements/BoxPlot.svelte
